requirements.txt
flask==3.1.2
flask-jwt-extended==4.7.1
python-dotenv==1.2.1
argon2-cffi==25.1.0
cryptography==42.0.8
SQLAlchemy==2.0.36
PyMySQL==1.1.1
flask-cors==4.0.0
